ppc/pnv: Add LPC controller and hook it up with a UART and RTC
This adds a model of the POWER8 LPC controller. It is then used
by the PowerNV code to attach a UART and RTC, which, with the right
version of OPAL firmware, will provide a working console.

This version of the LPC controller model doesn't yet implement
support for the SerIRQ deserializer present in the Naples version
of the chip though some preliminary work is there.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
[clg: updated for qemu-2.7
      ported on new sPowerNVMachineState which was merged with PnvSystem ]
Signed-off-by: Cédric Le Goater <clg@kaod.org>