ppc/pnv: don't populate device tree with serial devices
This is causing issues in the kernel running on the mainline skiboot,
in which the device tree is (brutally) fixed by adding a serial
device. For some reason, it generates a bogus node for the serial irq.

So let's remove the update of the device tree for serial device in
qemu for the moment until mainline choose a direction.

Signed-off-by: Cédric Le Goater <clg@kaod.org>