spapr: remove the XICS header from the machine and device models
The XICS routines are now completely hidden behind the sPAPR IRQ
interface. There is no need to include the header files anymore.

The only remaining direct use of XICS is done by the sPAPR cores which
need to create the IRQ presenter. This part will be removed when the
XIVE backend is introduced.

Signed-off-by: Cédric Le Goater <clg@kaod.org>