ppc/xive: introduce the XiveRouter model
The XiveRouter models the second sub-engine of the overall XIVE
architecture : the Interrupt Virtualization Routing Engine (IVRE).

The IVRE handles event notifications of the IVSE through MMIO stores
and performs the interrupt routing process. For this purpose, it uses
a set of table stored in system memory, the first of which being the
Interrupt Virtualization entries (IVE) table.

The IVT associates an interrupt source number with an Event Queue
Descriptor which will be used in a second phase of the routing process
to identify a notification target.

The XiveRouter is an abstract class which needs to be inherited from to
define a storage for the IVT, and other upcoming tables.

Signed-off-by: Cédric Le Goater <clg@kaod.org>