ppc/xive: notify the CPU when the interrupt priority is more privileged
After the event was pushed in the Xive EQ, the IVPE raises the bit
corresponding to the priority of the pending interrupt in the register
IBP (Interrupt Pending Buffer) to indicate there is an event pending
in one of the 8 priority queues. The Pending Interrupt Priority
Register (PIPR) is also updated using the IPB. This register represent
the priority of the most favored pending notification.

The PIPR is then compared to the the Current Processor Priority
Register (CPPR). If it is more favored (numerically less than), the
CPU interrupt line is raised and the EO bit of the Notification Source
Register (NSR) is updated to notify the presence of an exception for
the O/S. The check needs to be done whenever the PIPR or the CPPR are
changed.

The O/S acknowledges the interrupt with a special load in the Thread
Interrupt Management Area. If the EO bit of the NSR is set, the CPPR
takes the value of PIPR. The bit number in the IBP corresponding to
the priority of the pending interrupt is reseted and so is the EO bit
of the NSR.

Signed-off-by: Cédric Le Goater <clg@kaod.org>