spapr/xive: introduce a XIVE interrupt controller
sPAPRXive models the XIVE interrupt controller of the sPAPR machine.
It inherits from the XiveRouter and provisions storage for the routing
tables :

  - Interrupt Virtualization Entries (IVE)
  - Event Queue Descriptor (EQD)

sPAPRXive incorporates an internal XiveSource for the IPIs and the
virtual device interrupts of the guest. This model is consistent with
XIVE architecture which also incorporates an internal IVSE for IPIs
and accelerator interrupts in the IVRE sub-engine.

sPAPRXive exports two memory regions, one for the ESB trigger and
management pages used to control the sources and one for the TIMA
pages. They are mapped by default at the addresses found on chip 0 of
a real system. This is consistent with the XIVE architecture which
defines a Virtualization Controller BAR for the internal IVSE ESB
pages and a Thread Managment BAR for the TIMA.

Signed-off-by: Cédric Le Goater <clg@kaod.org>